[FEAT] Product-Image 보상 트랜잭션 재구현 및 Retry가 제대로 작동하도록 수정
1. DLT 토픽을 yml 파일에 등록
2. DLT를 처리하는 Handler 추가
3. KafkaListener에서 예외 발생 시 우선적으로 재시도를 하고, 그래도 안 될 시 fail 이벤트를 발송하도록 수정
  - 기존에는 KafkaErrorHandler가 구현되어 있었지만 catch로 예외를 다 잡아두어 이를 활용하지는 못함
4. 기존 보상 시 트랜잭션을 새로 시작하지 않던 문제 수정
  - ObjectProvider로 Self Injection 문제 해결
  - 보상 트랜잭션을 오케스트레이터 내부에서 그대로 사용
  - 그 대신, 새로운 트랜잭션을 생성하도록 전파 수준 수정

+++ src/main/java/io/devground/dbay/domain/image/infra/kafka/ImageKafkaListener.java
@@ -13,6 +13,9 @@ import io.devground.core.event.product.ProductImagesDeleteEvent;
 import io.devground.core.event.product.ProductImagesPushEvent;
 import io.devground.core.event.vo.EventType;
 import io.devground.core.model.vo.ImageType;
+import io.devground.dbay.common.saga.entity.Saga;
+import io.devground.dbay.common.saga.service.SagaService;
+import io.devground.dbay.common.saga.vo.SagaStep;
 import io.devground.dbay.domain.image.service.ImageService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -31,6 +34,7 @@ public class ImageKafkaListener {
 
 	private final ImageService imageService;
 	private final ImageKafkaProducer imageKafkaProducer;
+	private final SagaService sagaService;
 
 	@KafkaHandler
 	public void handleProductImagePush(ProductImagesPushEvent event) {
@@ -39,22 +43,22 @@ public class ImageKafkaListener {
 		String referenceCode = event.referenceCode();
 		ImageType imageType = event.imageType();
 
-		try {
-			String thumbnailUrl = imageService.saveImages(event.imageType(), event.referenceCode(), event.imageUrls());
+		Saga saga = sagaService.getSaga(sagaId);
+		SagaStep step = saga.getCurrentStep();
 
-			log.error("상품 이미지 등록 성공 - SagaId: {}, ProductCode: {}", event.sagaId(), event.referenceCode());
+		if (step.ordinal() >= SagaStep.IMAGE_DB_SAVE.ordinal()) {
+			log.warn("이미 종료된 이미지 등록 Saga - SagaId: {}, ProductCode: {}, Step: {}",
+				sagaId, referenceCode, step);
 
-			imageKafkaProducer.publishImageProcessed(
-				new ImageProcessedEvent(sagaId, imageType, referenceCode, EventType.PUSH, thumbnailUrl, true, null));
-
-		} catch (Exception e) {
-			log.error(
-				"상품 이미지 등록 실패 - SagaId: {}, ProductCode: {}, Exception: ", event.sagaId(), event.referenceCode(), e
-			);
-
-			imageKafkaProducer.publishImageProcessed(
-				new ImageProcessedEvent(sagaId, imageType, referenceCode, EventType.PUSH, null, false, e.getMessage()));
+			return;
 		}
+
+		String thumbnailUrl = imageService.saveImages(imageType, referenceCode, event.imageUrls());
+
+		log.info("상품 이미지 등록 성공 - SagaId: {}, ProductCode: {}", sagaId, referenceCode);
+
+		imageKafkaProducer.publishImageProcessed(
+			new ImageProcessedEvent(sagaId, imageType, referenceCode, EventType.PUSH, thumbnailUrl, true, null));
 	}
 
 	@KafkaHandler
@@ -65,27 +69,30 @@ public class ImageKafkaListener {
 		ImageType imageType = event.imageType();
 		List<String> deleteUrls = event.deleteUrls();
 
-		try {
-			if (CollectionUtils.isEmpty(deleteUrls)) {
-				imageService.deleteImageByReferences(imageType, referenceCode);
+		Saga saga = sagaService.getSaga(sagaId);
+		SagaStep step = saga.getCurrentStep();
 
-				log.info("상품 전체 이미지 삭제 완료 - SagaId: {}, ProductCode: {}", sagaId, referenceCode);
-			} else {
-				imageService.deleteImagesByReferencesAndUrls(
-					imageType, referenceCode, deleteUrls
-				);
+		if (step.ordinal() >= SagaStep.IMAGE_DELETED.ordinal()) {
+			log.warn("이미 종료된 이미지 삭제 Saga - SagaId: {}, ProductCode: {}, Step: {}",
+				sagaId, referenceCode, step);
 
-				log.info("선택된 상품 이미지 삭제 완료 - SagaId: {}, ProductCode: {}, size: {}",
-					sagaId, referenceCode, deleteUrls.size());
-			}
-
-			imageKafkaProducer.publishImageProcessed(
-				new ImageProcessedEvent(sagaId, imageType, referenceCode, EventType.DELETE, null, true, null));
-		} catch (Exception e) {
-			log.error("상품 이미지 삭제 실패 - SagaId: {}, ProductCode: {}, Exception: ", sagaId, referenceCode, e);
-
-			imageKafkaProducer.publishImageProcessed(
-				new ImageProcessedEvent(sagaId, imageType, referenceCode, EventType.DELETE, null, false, e.getMessage()));
+			return;
 		}
+
+		if (CollectionUtils.isEmpty(deleteUrls)) {
+			imageService.deleteImageByReferences(imageType, referenceCode);
+
+			log.info("상품 전체 이미지 삭제 완료 - SagaId: {}, ProductCode: {}", sagaId, referenceCode);
+		} else {
+			imageService.deleteImagesByReferencesAndUrls(
+				imageType, referenceCode, deleteUrls
+			);
+
+			log.info("선택된 상품 이미지 삭제 완료 - SagaId: {}, ProductCode: {}, size: {}",
+				sagaId, referenceCode, deleteUrls.size());
+		}
+
+		imageKafkaProducer.publishImageProcessed(
+			new ImageProcessedEvent(sagaId, imageType, referenceCode, EventType.DELETE, null, true, null));
 	}
 }
